Alternate approach: Classloader & Policy
